Add ability to update API from a backup file
Add new feature to update/restore API from the temporary backup file in case the script is unable to update API for the latest data. It will be restored from the most recent successful API update. So even if the twitter server is down, or API couldn't be updated for any other reason, the script will not crash, it will use the most recent API endpoints/data from the backup file.